Prog: oop

+++ Programacion/00 - rough testing/src/ej_ricardo_empleados/JefeAlmacen.java
@@ -4,7 +4,7 @@ package ej_ricardo_empleados;
 public class JefeAlmacen extends Empleado {
   int storehouse_id;
 
-  public JefeAlmacen(String name, String birthDate, String post, float baseSalary) {
+  public JefeAlmacen(String name, String birthDate, String post, double baseSalary) {
     super(name, birthDate, post, 3, baseSalary);
   }
 }
